src/auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.0.1"
description = "demo auth microservice"
authors = ["Lucas Lucyk <llucyk@grupospec.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "auth"}, {include = "cli"}, {include = "api"}]

[tool.poetry.dependencies]
python = "^3.8"
sa-modelcrud = "^0.3.2"
pydantic-settings = "^2.0.3"
pydantic = {extras = ["email"], version = "^2.4.2"}
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
cryptography = "^41.0.5"
aiosqlite = "^0.19.0"
alembic = "^1.12.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
typer = {extras = ["all"], version = "^0.9.0"}


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
